Ocana Update
I fixed some spelling and grammar. I didn't look at the "Model intuition" section. I added some of my ideas to the "How to update", "How to ground truth", and "How to not abuse the Lester model" sections. I added some quotes from Lester et al. to the "How to not abuse the Lester model" section. I wasn't sure what to add to the "How to not abuse the Lester model" section. We want a tool to manage individual lakes, which is abusing the Lester model.